Refactor & Redesign minor refactoring, movving files between folders.
